Move is_static to utils
